icons/requirements-icon.txt
# Requirements for icon generation
Pillow>=10.0.0 